asmtool: Prevent long overflow on lcs_matrix
The length of lcs_matrix's m_matrix is defined based on the number of elements
of the compared entities. Thus, if the multiplication of the number of elements
on both entities exceed what can be held in a long variable, this causes an
overflow and the matrix is allocated with a wrong length.

Fixing this would require using a variable type with more bits while defining
the matrix length, but then the allocated memory would exceed possible limits
and crash execution.

Add a check for multiplication overflow and, if this is the case, skip it and
drop a warning informing that the entity cannot be handled/diff'ed.

The bug can be reproduced by comparing this file (gcc-generated) with itself:
* http://www.livewire.com.br/suse/asmtool/mmu_pv.s.klp

Signed-off-by: Joao Moreira <jmoreira@suse.de>
Signed-off-by: Joerg Roedel <jroedel@suse.de>